refactor: ConsoleReader 클래스의 패키지를 util로 변경
